improve robustness and type safety
- Remove .legend() call from quickstart example without color mapping
- Add .toDouble() casting for clamp() calls in legend examples
- Add defensive check for empty colorPalette to prevent division by zero
- Fix Flex constraints issue with LayoutBuilder and Flexible widgets
- Add overlay fallback for unbounded constraint scenarios

+++ example/lib/graphs/legend_example.dart
@@ -34,7 +34,7 @@ Widget buildLegendExampleTab(ChartTheme currentTheme,
                 .data(data)
                 .mapping(x: 'quarter', y: 'revenue', color: 'product')
                 .geomBar(
-                    width: sliderValue.clamp(0.1, 1.0),
+                    width: sliderValue.clamp(0.1, 1.0).toDouble(),
                     style: BarStyle.grouped,
                     alpha: 0.9)
                 .scaleXOrdinal()
@@ -63,7 +63,7 @@ Widget buildLegendExampleTab(ChartTheme currentTheme,
                 .data(data)
                 .mapping(x: 'quarter', y: 'revenue', color: 'product')
                 .geomBar(
-                    width: sliderValue.clamp(0.1, 1.0),
+                    width: sliderValue.clamp(0.1, 1.0).toDouble(),
                     style: BarStyle.grouped,
                     alpha: 0.9)
                 .scaleXOrdinal()
@@ -93,7 +93,7 @@ Widget buildLegendExampleTab(ChartTheme currentTheme,
                 .data(data)
                 .mapping(x: 'quarter', y: 'revenue', color: 'product')
                 .geomBar(
-                    width: sliderValue.clamp(0.1, 1.0),
+                    width: sliderValue.clamp(0.1, 1.0).toDouble(),
                     style: BarStyle.grouped,
                     alpha: 0.9)
                 .scaleXOrdinal()
@@ -133,7 +133,7 @@ Widget buildLegendExampleTab(ChartTheme currentTheme,
                 .data(data)
                 .mapping(x: 'quarter', y: 'revenue', color: 'product')
                 .geomBar(
-                    width: sliderValue.clamp(0.1, 1.0),
+                    width: sliderValue.clamp(0.1, 1.0).toDouble(),
                     style: BarStyle.grouped,
                     alpha: 0.9)
                 .scaleXOrdinal()
